move isSubdirectory to views
