+++ views/utils/tools.es
@@ -161,33 +161,6 @@ export function compareUpdate(prevState, newState, depth=1) {
   })
   return prevState
 }
-/* TEST
-function test(a, b, d) {
-  const c = compareUpdate(a, b, d)
-  console.log(c !== a, c)
-}
-
-test(2, 2)
-// false 2
-test({1:'a'},{2:'b'})
-// true {"1":"a","2":"b"}
-test({1:'a'},{1:'b'})
-// true {"1":"b"}
-test({1:'a'},{1:'a'})
-// false {"1":"a"}
-test({1:{1:2}},{1:{1:2}})
-// false {"1":{"1":2}}
-test({1:{1:[], 2:['g']}},{1:{1:[]}})
-// true {"1":{"1":[]}}
-test({1:{1:[], 2:['g']}},{1:{1:[]}}, 2)
-// false {"1":{"1":[],"2":["g"]}}
-
-let a=[]
-a[1] = {1:2}
-test([{1:1}],a)
-// true [{"1":1},{"1":2}]
-
-*/
 
 function pad(n) {
   return n < 10 ? `0${n}` : n
@@ -223,4 +196,12 @@ export const fileUrl = (str = '') => {
     pathName = '/' + pathName
   }
   return 'file://' + pathName
+}
+
+// check if dir is a subdirectory of parent,
+// if parent and dir are the same, also returns true
+export const isSubdirectory = (parent, dir) => {
+  const relative = path.relative(parent, dir)
+  return !relative ||
+    (!relative.startsWith('..') && !path.isAbsolute(relative))
 }
